Add native function call support.
Syntax proposal that introduces the "native" keyword in order to specify that a function body
is bind to a native one.

Native function might be implemented with the JNI API. All Golo functions might be declared as
"native" including closures and functions declared in a pimp. Native
name must be specified  for the closures in order to have more friendly name in native code.