Changing from "pimps" to "class augmentations"
This change is being made to accomodate cultural differences around the world.

The new keyword is 'augment' instead of 'pimp'. The 'pimp' keyword will continue to remain valid in
the parser for some time, although its existence has been removed from the documentation and the
terminology used in the source code.

Signed-off-by: Julien Ponge <julien.ponge@insa-lyon.fr>