Fix to prevent completion from confusing the golo command with the golo subcommand.
Trying to use autocompletion for 'golo golo' would result in auto completion for the golo subcommand if you just tried to do auto completion for the golo command. COMPWORDS includes the top level command, so we just need to look for a special word from the 2nd indice instead of the first.