Reducing the languages bridges factor
Computations of the GCD now have a repeat count factor to increase the time spent in language
realms. Results are being accumulated in a way that shall be free of dead code elimination, with the
loop bound being a mutable field parameter.

Measurement has been change from a throughput to a single-shot batched time measure.

The dominant factor is now how languages deal with boolean and number types for computations and
loops/branches logic.