AdapterFabric-based asFunctionalInterface
This defeats the limitations of LambdaMetaFactory which is caller-sensitive. I also tried
sun.reflect.Reflection.getCallerClass() to hook the LambdaMetaFactory lookup into the caller class,
but I encountered errors.

I also tried to generalize to an all AdapterFabric-based implementation, but it would get
ClassNotFoundException when doing adaptation on parameters.

This implementation is a compromise:

- LambdaMetaFactory get the nod for argument adaptation, and
- AdapterFabric powers asFunctionalInterface to bypass caller-sensitive checks.