Boxing of MethodHandle into FunctionReference
This is a backwards-incompatible change.
gololang.FunctionReference is a new class to box MethodHandle references.

The rationale for boxing function / closure references is that of attaching state to method handles,
which is not possible since MethodHandle is a final class, and is motivated by the future introduction
of named parameters.

At this stage the boxing does not add any new state. This will done in future works, but the nice
takeaway is that we now have an indirection in place.

Plugging FunctionReference in place of MethodHandle required lots of modifications. At this stage the
code passes all tests, but we will still need to:

1. test on larger code bases, and
2. port previous MH augmentations into the new FunctionReference class, and
3. update the documentation (user guide, {java,golo}docs).

FunctionReference delegates most methods of MethodHandle.
I did not delegate the invoke{...}(Object... args) methods, as varargs delegation does not work well.
Instead the idea is that calls through methods need to go through the MH unboxing using the handle()
method, which prevents array wrapping issues due to varargs over varargs.

Note that ClosureCallSupport still builds a method handle chain with inline-caching, hence performance
is not expected to suffer for regular calls such as:

	let f = |n| -> n + 1
	println(f(1))

Calls through methods are likely to remain as fast, since the unboxing indirection is likely to cause
handle() to be inlined:

	let f = |n| -> n + 1
	println(f: handle(): invoke(1)) # instead of f: invoke(1) in previous Golo versions

Signed-off-by: Julien Ponge <julien.ponge@insa-lyon.fr>