[471709] Exception when comparing numbers
The implementation of comparison operators was done using `compareTo`,
which fails when comparing two numbers of different classes.

The bug is fixed by using `Double.compare` on the two numbers
`doubleValue` when both values are instances of `Number`, falling back
to `compareTo` otherwise.

The same is done for equality, since until now, golo considered
`1 != 1.0`. When the two values are numbers, the double value is
compared by value, `equals` is called otherwise.

Bug: https://bugs.eclipse.org/bugs/show_bug.cgi?id=471709