Alternative implementation
Use the `generate_math.rb` script to generate type specific operators
with explicit cast and guard.
The operators semantic will then be the same as Java on primitive types,
falling back to `equals` and `compareTo`.
This allows to use the binary operator inline cache.

Also use `java.util.Objects.equals` for `==` and `!=` implementation.