improve the Maven publication in the gradle build
Instead of relying on a profile to select the Maven repository when publishing, configure the build to generate one task per target Maven repository. This allows publishing to multiple repositories at once, and makes it easier to know how to publish by simply executing `gradlew tasks`, and have a list of all publication tasks available.

Signed-off-by: jbnizet <jb@ninja-squad.com>