Operator call sites with catchException instead of a guardWithTest
This changes OperatorSupport to leverage a ClassCastException to rewrite call sites instead of
using a guard with test. With stable call sites this removes 2 instanceof checks and appears
to slightly improve performance. This is not a claim though, further performance analysis is
required.

The guardWithTest construct is still present for correctness: when an argument of a binary
operator is a String, then it allows detecting when the call site should rewrite back to a
numeric operation. Indeed, when the call site is concatenating String and (any) Object, then
there are few cases where a ClassCastException can be thrown to invalidate the call site:

    Integer + Integer -> catchException
    String  + Integer -> guardWithTest, invalidated by a CCE
    String  + String  -> stable
    Integer + Long    -> catchException, invalidated by arguemnt checks
    (...)