FIX #320: Lookup augmentations in call stack
When an augmentation is defined, it is only visible if the module using
it has imported the module defining it. This limits the ability to
create generic functions using any augmented object.

The solution is to look up in the call stack if a complying augmentation
had been imported in the calling module.