Refactor method and augmentation resolution
Change the way (augmented) methods are resolved at runtime, to deal
with the priorities discussed in #326.

Move methods common between augmentation and regular method resolution
into a abstract superclass.

Refactor method type matching, to better reuse and move it to the
MethodInvocation object.

FIX #275

Also some refactoring on tests.

FIX: empty augmentation

Not filled as an issue, but a class was generated to hold augment
functions, even if there was no functions. This happened when using
named augmentation, since a single IR node is created for declarations
such as

    augment Foo {
      function plop = |this| -> "plop"
    }

and

    augment Foo with Plop

In the first case a class `module$Foo` is created to hold the `plop`
function. It was also created in the second case, but was empty.