FIX #335: Named arguments and anonymous calls
There is a issue when doing anonymous calls, i.e. calls directly applied
on references, on lambda, on partialized functions, or on functions
returned by other functions, when using named arguments.

The issue is due to an incorrect flagging in the AST to IR conversion:
the anonymous function invocations were not flagged as using named
arguments. Therefore, the magic behind did not happened, which caused
the arguments not being stacked, and thus a weired runtime error.

(Also some space cleaning)