Check command
Adds a `check` command to the `golo` binary. This command does the same
tasks as the `compile` command, namely parse input files and run the
compile time IR verifications, but does not produce result files.

This command can be useful to integrate golo with editor plugins such as
Vim Syntastic.

FIX #324