Refactor anonymous call chains
This FIX #385: Heterogeneous anonymous invocation

There is currently two ways to represent an anonymous function invocation
in the IR: as a list of anonymous invocations in the
`ir.AbstractInvocation` class, or as a `ir.BinaryOperation` with a
`OperatorType.ANON_CALL` type. Although the bytecode generated in both
cases is the same and the difference is transparent to the end user,
this leads to a more complex IR that could be simplified.

Only the operator version is kept, since:
- some weired call/invocation sequences don't parse,
- the IR would be more consistent and simple (though complex sequences
  will be deeply nested),
- it would make implementing #363 (Guarded call operator) and other
  future improvement more natural.

The grammar and IR are simplified to remove the anonymous invocation
sequences `anonymousFunctionInvocation`, and replace it with a
`BinaryOperation` of type `ANON_CALL`.

Remove the `AnonymousFunctionInvocation` from `FunctionInvocation` and
let the `InvocationExpression` deal with anon calls.
Dispatch the change to decorators.

Decorators are also adapted to accept a simple closure reference.