FIX #248: local type must be fully qualified
This PR fixes the #248 bug.
When defining new types (`struct` or `union`), the classes are compiled
in a `[moduleName].types` sub-package, that is not directly accessible
from the defining module. As a result, the class can't be accessed
directly without fully qualifying it or importing the `types`
sub-package.

For instance, the following code fails

    module foo

    struct Foo = {a}

    function main = |args| {
      println(Foo.class)
    }

with a `BootstrapMethodError` since `Foo.types` is not looked up.

The proposed fix is quite simple: it adds an
`import [currentModule.types` as the last `import` statement to every module.