Integrate comments and clean code
Update the doc and tests to reflect the fact that since we now generate
factory functions for union values in the module that defines them (for
default values), we no longer need to implicitly import the type or
prefix factories with the union name.

Also adds named parameters to union methods.