Package index documentation
Adds package wide documentation (like `package-info.java`) by using
empty modules (no class generated, only doc) or `README.md` in the same
directory. The package doc links to all modules in the same package.
Such a doc is generated even if no info is provided, only listing
contained modules.

Even when the module is not empty, if modules exists in a "subnamespace"
(e.g. `my.module` and `my.module.sub`), the upper module links to the
others.

Moreover, a description of the module (from the first sentence of the
module golodoc) is included in the index.