Fix a bug on self reference in closures
When a closure references itself elsewhere than in a recursive call,
an `Uninitialized reference` error is raised. For instance, in

```golo
function test = {
  let fun = |x| -> match {
    when x is null then fun
    otherwise x
  }
  return fun(null)
}
```

the null case fails. In the generated function `__$$_sugar_closure_X`,
an assignment to `fun` of the closure reference is injected, but it was
mistakenly not defined as a declaring one, which fool the
`LocalReferenceAssignmentAndVerificationVisitor` in thinking the
reference is not initialized.