Add l10n to cli, error messages, and exceptions
The error messages and exceptions, as well as the CLI options are
localized.

A class to ease the printing of errors, infos and warnings is also
added, allowing to print easily localized messages on stderr, using
colors if possible.
This class is in the `gololang` package to allow users to reuse it in
golo programs.