Implicit import of the current package
In a module `foo.bar.Baz`, we add an implicit `import foo.bar`, such
that modules in the same package are directly visible.
This mimic the Java behavior in which classes in the same package are
directly visible without importing them.

For instance, given

    module foo.bar.Plop

    function daPlop = -> "Plop"

the module `foo.bar.Baz` can directly call `Plop.daPlop()`.