Fix #380 on variable scope
As discussed in the issue thread, the expected behavior is not that
clear. The current one mimic the java one, it is indeed a realy bad idea
to redefine/shadow an existing variable in a more local scope.

This fix just display the correct error message that we are redefining a
variable.