Lazy project type initialization
To improve the `golo new` command, project type initializers were
introduced. However, they were stored in a static map, and used golo
template for the files to generate. Therefore, a template compilation
was triggered for every golo run, even when an other command was used.

The initializer are made lazy to prevent this behavior.