Fix a bug in function composition
When composing a variadic function, the collector status is lost, thus
the resulting function can't be used as varargs.

For instance:

```golo
let f = (|a...| -> a: toString()): andThen(|a| -> a)

f("a", "b")
f: invoke("a", "b")
```

Both calls fail with a `WrongMethodTypeException`. A solution is to add
a `: asVarargsCollector()` call, but it is counter intuitive: the
composed function should directly accept the same arguments as the first
function and return the same type as the second one.

This is fixed by converting the composed function into a varargs
collector if the original one was.

The same issue occurred when binding parameters on varargs functions.

A last bug fixed here occurred when composing java functions whose types
do not match exactly even if compatible
(e.g. composing `A -> String` with `Object -> B` to get `A -> B`)
The types of both functions are now adapted to be `Object`. However, a
`ClassCastException` can be raised on invocation if types are not
actually compatible.

Another change introduced is to accept to compose with a parameterless
function, as in:

```golo
let f = (|a|{println("got " + a)}): andThen(-> 42)
println(f("answer"))
```

This used to raise an `IllegalArgumentException`. A solution was to use
`: andThen(|ignored| -> 42)` instead, which is cumbersome and prevents
to use existing functions.

The result of the first function is now ignored. This is useful to
sequence functions with side effects, that return null/void. The
previous example prints

    got answer
    42