Better way to make tests pass
Define a thread local variable to hold the current thread classloader.

When running from golo, the current thread classloader will be a
GoloClassLoader instance, so it is used as is.
The issue is that when running from Java, the current classloader is not
a golo one, so it was wrapped by `Runtime.classLoader()`. We
therefore got a different one for each call.
Keeping it in a thread local singleton resolve this issue.